requirements.txt
flask
Flask-SQLAlchemy
Flask-Admin
pymysql
flsk_stp
python-telegram-bot
werkzeug.utils